pyproject.toml
[project]
name = "nhl"
version = "0.1.0"
description = "NHL Scores and Schedule Client for Raspberry Pi Zero"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.25.0",
    "inky>=2.2.1",
    "python-dotenv>=0.19.0",
]
